Make plutus-pab's local cluster in it's own library which can be run using user defined builtin contracts.
* This commit is part of an more general issue to inverse the dependency between plutus-pab and plutus-use-cases projects.

* Added TODO comments in plutus-pab's file on was needs to be done.

* Changed the name the datatype `ExampleContracts` to singular `ContractExample` which makes more sense.
