+++ plutus-pab/examples/Main.hs
@@ -4,10 +4,10 @@ module Main
     ( main
     ) where
 
-import ContractExample (ExampleContracts)
+import ContractExample (ContractExample)
 import Plutus.PAB.Effects.Contract.Builtin qualified as Builtin
 import Plutus.PAB.Run (runWith)
 
 main :: IO ()
 main = do
-    runWith (Builtin.handleBuiltin @ExampleContracts)
+    runWith (Builtin.handleBuiltin @ContractExample)
